Revert "Remove NotImplementedErrors by already existing implementation"
This reverts commit 51f723e36e973c21414a92b99fd2680e9db9d3fb.
Causes problems with unittests. The function load_audio_file tries first loading the data with
the function I changed and if an error is thrown (i.e. if rescaling/resampling is required) it
falls back to ffmpeg, where the resulting signal array differs slightly. This breaks some
unittests and doctests.